requirements.txt
streamlit
snowflake-connector-python
pandas
numpy
requests
python-dotenv
